fix(us&cs)/fix CORS

+++ services/core-service/src/domain/services/auth.service.ts
@@ -56,7 +56,9 @@ export class AuthService {
 			lastUse: new Date().toISOString(),
 		});
 
-		await this.registerWhitelistRepository.deleteRegisterWhitelistById(whitelistId);
+		if (!databaseEmpty) {
+			await this.registerWhitelistRepository.deleteRegisterWhitelistById(whitelistId);
+		}
 
 		return { user, session: sessionToken };
 	}
